Made some changes
Have to take note that Pavi and Kevin's will most likely be better. I have the Activity Indicator fixed a bit. Pavi should have the icon for the offline/online and other stuff. Kevin has a lot of his own. I dont know if you wanna use this one because I do not have all of the changes.